// src/shared/receipts/types.ts
export type ReceiptTemplateModel = {
  documentId?: string | null
  documentNumber?: string | null
  documentType?: string | null
  isOnline?: string | null
  isFiscalized?: string | null
  companyName?: string | null
  companyMobile?: string | null
  companyTin?: string | null
  companyVrn?: string | null
  companySerial?: string | null
  companyTaxOffice?: string | null
  receiptNumber?: string | null
  receiptZNumber?: string | null
  receiptDate?: string | null
  receiptTime?: string | null
  receiptInternalData?: string | null
  receiptSignature?: string | null
  fiscalVerificationCode?: string | null
  fiscalQrCodeData?: string | null
}